Saha and Peebles now working.
Get a plot of electron fraction that matches previous results (Callin 2006).
However, when zooming in on the region where Peebles is taking over for Saha,
there is an artifact: one point with slightly higher value than it should. Have
found no way around this yet.

Have also written a function that splines the electron density. Tested and
working.