Tight coupling and post-tight coupling working
Runtime is currently at 2 minutes using Numba for the ODE functions.
Tried to speed it up by making functions for the non-ODE's in the TC regime, and
rather call these in the k-loop instead of solving the ODE system with these
non-ODEs. However, it had no effect, and both versions are running at 2 minutes.